feat(editor): add game API for saving
